+++ src/rgw/CMakeLists.txt
@@ -38,7 +38,6 @@ set(librgw_common_srcs
   services/svc_otp.cc
   services/svc_quota.cc
   services/svc_sync_modules.cc
-  services/svc_rados.cc
   services/svc_sys_obj.cc
   services/svc_sys_obj_cache.cc
   services/svc_sys_obj_core.cc
@@ -46,6 +45,7 @@ set(librgw_common_srcs
   services/svc_user.cc
   services/svc_user_rados.cc
   services/svc_zone.cc
+  services/rados/svc_rados.cc
   services/svc_zone_utils.cc
   rgw_service.cc
   rgw_acl.cc
