+++ src/rgw/rgw_object_expirer_core.cc
@@ -32,7 +32,7 @@
 #include "rgw_zone.h"
 #include "rgw_sal_rados.h"
 
-#include "services/svc_rados.h"
+#include "services/rados/svc_rados.h"
 #include "services/svc_zone.h"
 #include "services/svc_sys_obj.h"
 #include "services/svc_bi_rados.h"
