+++ src/rgw/rgw_putobj_processor.h
@@ -18,7 +18,7 @@
 #include <optional>
 
 #include "rgw_putobj.h"
-#include "services/svc_rados.h"
+#include "services/rados/svc_rados.h"
 #include "services/svc_tier_rados.h"
 #include "rgw_sal.h"
 #include "rgw_obj_manifest.h"
